backend/requirements.txt
# ==================== Core Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ==================== Database ====================
sqlalchemy==2.0.23
asyncpg==0.29.0          # PostgreSQL async driver
alembic==1.13.0          # Database migrations
psycopg2-binary==2.9.9   # PostgreSQL sync driver (for Alembic)

# ==================== Authentication ====================
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4            # Password hashing
bcrypt==4.1.1
google-auth==2.25.2               # Google OAuth
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0

# ==================== Validation & Settings ====================
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# ==================== HTTP Client ====================
httpx==0.25.2
aiohttp==3.9.1

# ==================== Caching ====================
aiocache==0.12.2  # Async caching library with Redis backend
redis[hiredis]==5.0.1  # Redis async client (replaces aioredis for Python 3.12 compatibility)

# ==================== AI Providers ====================
google-generativeai==0.3.1  # Google Gemini
openai==1.6.1               # OpenAI (GPT) - 추후 사용
elevenlabs==0.2.26          # ElevenLabs TTS

# ==================== File Processing ====================
pillow==10.1.0              # Image processing
python-magic==0.4.27        # File type detection
aiofiles==23.2.1            # Async file operations
boto3==1.34.0               # AWS SDK

# ==================== Testing ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2               # Test client

# ==================== Code Quality ====================
ruff==0.1.8                 # Linter (fast alternative to flake8)
black==23.12.1              # Code formatter
isort==5.13.2               # Import sorter
mypy==1.7.1                 # Type checker

# ==================== Utilities ====================
python-dateutil==2.8.2
pytz==2023.3
